+++ cmd/ged/accountcmd.go
@@ -41,7 +41,7 @@ var (
 			{
 				Action: importWallet,
 				Name:   "import",
-				Usage:  "import earthdollar presale wallet",
+				Usage:  "import ethereum presale wallet",
 			},
 		},
 		Description: `
@@ -76,7 +76,7 @@ Note that exporting your key in unencrypted format is NOT supported.
 
 Keys are stored under <DATADIR>/keystore.
 It is safe to transfer the entire directory or the individual keys therein
-between earthdollar nodes by simply copying.
+between ethereum nodes by simply copying.
 Make sure you backup your keys regularly.
 
 In order to use your account to send transactions, you need to unlock them using
@@ -100,7 +100,7 @@ And finally. DO NOT FORGET YOUR PASSWORD.
 				Usage:  "Create a new account",
 				Description: `
 
-    ged account new
+    geth account new
 
 Creates a new account. Prints the address.
 
@@ -110,7 +110,7 @@ You must remember this passphrase to unlock your account in the future.
 
 For non-interactive use the passphrase can be specified with the --password flag:
 
-    earthdollar --password <passwordfile> account new
+    ethereum --password <passwordfile> account new
 
 Note, this is meant to be used for testing only, it is a bad idea to save your
 password to file or expose in any other way.
@@ -122,7 +122,7 @@ password to file or expose in any other way.
 				Usage:  "Update an existing account",
 				Description: `
 
-    ged account update <address>
+    geth account update <address>
 
 Update an existing account.
 
@@ -134,7 +134,7 @@ format to the newest format or change the password for an account.
 
 For non-interactive use the passphrase can be specified with the --password flag:
 
-    ged --password <passwordfile> account update <address>
+    geth --password <passwordfile> account update <address>
 
 Since only one password can be given, only format update can be performed,
 changing your password is only possible interactively.
@@ -146,7 +146,7 @@ changing your password is only possible interactively.
 				Usage:  "Import a private key into a new account",
 				Description: `
 
-    ged account import <keyfile>
+    geth account import <keyfile>
 
 Imports an unencrypted private key from <keyfile> and creates a new account.
 Prints the address.
@@ -159,10 +159,10 @@ You must remember this passphrase to unlock your account in the future.
 
 For non-interactive use the passphrase can be specified with the -password flag:
 
-    ged --password <passwordfile> account import <keyfile>
+    geth --password <passwordfile> account import <keyfile>
 
 Note:
-As you can directly copy your encrypted accounts to another earthdollar instance,
+As you can directly copy your encrypted accounts to another ethereum instance,
 this import mechanism is not needed when you transfer an account between
 nodes.
 					`,
@@ -173,7 +173,7 @@ nodes.
 				Usage:  "Build persistent account index",
 				Description: `
 
-    ged --index-accounts account index
+    geth --index-accounts account index
 
 Create keystore directory index cache database (keystore/accounts.db). Relevant for use with large amounts of key files (>10,000).
 
@@ -192,7 +192,7 @@ It indexes all key files from keystore/* (non-recursively).
 func accountIndex(ctx *cli.Context) error {
 	n := aliasableName(AccountsIndexFlag.Name, ctx)
 	if !ctx.GlobalBool(n) {
-		log.Fatalf("Use: $ ged --%v account index\n (missing '%v' flag)", n, n)
+		log.Fatalf("Use: $ geth --%v account index\n (missing '%v' flag)", n, n)
 	}
 	am := MakeAccountManager(ctx)
 	errs := am.BuildIndexDB()