+++ cmd/ged/consolecmd.go
@@ -31,11 +31,11 @@ var (
 	consoleCommand = cli.Command{
 		Action: localConsole,
 		Name:   "console",
-		Usage:  `Ged Console: interactive JavaScript environment`,
+		Usage:  `Geth Console: interactive JavaScript environment`,
 		Description: `
-The Ged console is an interactive shell for the JavaScript runtime environment
+The Geth console is an interactive shell for the JavaScript runtime environment
 which exposes a node admin interface as well as the Ðapp JavaScript API.
-See https://github.com/Tzunami/go-earthdollar/wiki/Javascipt-Console
+See https://github.com/ethereum/go-earthdollar/wiki/Javascipt-Console
 `,
 		Flags: []cli.Flag{
 			ExecFlag,
@@ -44,12 +44,12 @@ See https://github.com/Tzunami/go-earthdollar/wiki/Javascipt-Console
 	attachCommand = cli.Command{
 		Action: remoteConsole,
 		Name:   "attach",
-		Usage:  `Ged Console: interactive JavaScript environment (connect to node)`,
+		Usage:  `Geth Console: interactive JavaScript environment (connect to node)`,
 		Description: `
-The Ged console is an interactive shell for the JavaScript runtime environment
+The Geth console is an interactive shell for the JavaScript runtime environment
 which exposes a node admin interface as well as the Ðapp JavaScript API.
-See https://github.com/Tzunami/go-earthdollar/wiki/Javascipt-Console.
-This command allows to open a console on a running ged node.
+See https://github.com/ethereum/go-earthdollar/wiki/Javascipt-Console.
+This command allows to open a console on a running geth node.
 	`,
 		Flags: []cli.Flag{
 			ExecFlag,
@@ -58,15 +58,15 @@ This command allows to open a console on a running ged node.
 	javascriptCommand = cli.Command{
 		Action: ephemeralConsole,
 		Name:   "js",
-		Usage:  `Executes the given JavaScript files in the Ged JavaScript VM`,
+		Usage:  `Executes the given JavaScript files in the Geth JavaScript VM`,
 		Description: `
 The JavaScript VM exposes a node admin interface as well as the Ðapp
-JavaScript API. See https://github.com/Tzunami/go-earthdollar/wiki/Javascipt-Console
+JavaScript API. See https://github.com/ethereum/go-earthdollar/wiki/Javascipt-Console
 `,
 	}
 )
 
-// localConsole starts a new ged node, attaching a JavaScript console to it at the
+// localConsole starts a new geth node, attaching a JavaScript console to it at the
 // same time.
 func localConsole(ctx *cli.Context) error {
 	// Create and start the node based on the CLI flags
@@ -77,7 +77,7 @@ func localConsole(ctx *cli.Context) error {
 	// Attach to the newly started node and start the JavaScript console
 	client, err := node.Attach()
 	if err != nil {
-		log.Fatal("Failed to attach to the inproc ged: ", err)
+		log.Fatal("Failed to attach to the inproc geth: ", err)
 	}
 	config := console.Config{
 		DataDir: node.DataDir(),
@@ -112,10 +112,10 @@ func localConsole(ctx *cli.Context) error {
 	return nil
 }
 
-// remoteConsole will connect to a remote ged instance, attaching a JavaScript
+// remoteConsole will connect to a remote geth instance, attaching a JavaScript
 // console to it.
 func remoteConsole(ctx *cli.Context) error {
-	// Attach to a remotely running ged instance and start the JavaScript console
+	// Attach to a remotely running geth instance and start the JavaScript console
 	chainDir := MustMakeChainDataDir(ctx)
 	var uri = "ipc:" + node.DefaultIPCEndpoint(chainDir)
 	if ctx.Args().Present() {
@@ -123,7 +123,7 @@ func remoteConsole(ctx *cli.Context) error {
 	}
 	client, err := rpc.NewClient(uri)
 	if err != nil {
-		log.Fatal("attach to remote ged: ", err)
+		log.Fatal("attach to remote geth: ", err)
 	}
 
 	config := console.Config{
@@ -159,7 +159,7 @@ func remoteConsole(ctx *cli.Context) error {
 	return nil
 }
 
-// ephemeralConsole starts a new ged node, attaches an ephemeral JavaScript
+// ephemeralConsole starts a new geth node, attaches an ephemeral JavaScript
 // console to it, and each of the files specified as arguments and tears the
 // everything down.
 func ephemeralConsole(ctx *cli.Context) error {
@@ -171,7 +171,7 @@ func ephemeralConsole(ctx *cli.Context) error {
 	// Attach to the newly started node and start the JavaScript console
 	client, err := node.Attach()
 	if err != nil {
-		log.Fatal("Failed to attach to the inproc ged: ", err)
+		log.Fatal("Failed to attach to the inproc geth: ", err)
 	}
 	config := console.Config{
 		DataDir: node.DataDir(),