+++ cmd/ged/usage.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with go-earthdollar. If not, see <http://www.gnu.org/licenses/>.
 
-// Contains the ged command usage template and generator.
+// Contains the geth command usage template and generator.
 
 package main
 
@@ -116,7 +116,7 @@ var AppHelpFlagGroups = []flagGroup{
 			MinerThreadsFlag,
 			MiningGPUFlag,
 			AutoDAGFlag,
-			EarthbaseFlag,
+			EtherbaseFlag,
 			TargetGasLimitFlag,
 			GasPriceFlag,
 			ExtraDataFlag,